legal/AI/Ai_boon/requirements.txt
# =================================================================
# BoonAI 통합 법률 AI FastAPI 의존성
# 분류 + RAG + Ollama LLM (LangChain 없음)
# =================================================================

# FastAPI 프레임워크
fastapi>=0.109.0
uvicorn>=0.27.0

# 토치 및 트랜스포머 (분류 모델)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.10.0+cpu
transformers>=4.37.0

# RAG / 벡터 스토어 (LangChain 없이 직접 사용)
faiss-cpu>=1.7.4
sentence-transformers>=2.3.1

# HTTP 클라이언트 (Ollama API 호출)
httpx>=0.27.0

# 기타 의존성
pydantic>=2.5.3
numpy<2.0.0
